+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a
 Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.8.0 - 2021-04-26
+
+### Added
+
+- Added the `:read_only?` configuration flag for
+  `Spear.Connection.Configuration`
+    - this allows one to limit what the `Spear.Connection` will perform to
+      read-only operations such as reading streams
+
 ## 0.7.0 - 2021-04-24
 
 ### Added